requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12.0
django-bootstrap4>=22.3
requests>=2.25.1 
behave
django-behave
